Add files via upload
# Arduino Morse Identifier
This program is written in Visuino open wire format.

Morse code is manually entered into the TextValueMorse object. Use 1 for tone, 0 for silence. The letter 'A' (di-dah) would be encoded as 101110. A dot (1) an inter-character space (0) then a dash (111). For best results, use an inter-character space at the end of each character and use three zeros (000) for an inter-word space. So 'hi world', which sounds like this: di-di-di-dit  di-dit       di-dah-dah  dah-dah-dah  di-dah-dit  di-dah-di-dit  dah-di-dit, would be encoded as: 
```
  H        I   Sp  W          O            R         L           D
010101010001010000101110111001110111011100101110100010111010100011101010
```
Digital output pin 2 is used to key the transmitter
Morse code characters are output on pin 8. Hook an LED here to see the code.
Place a small capacitor (10uF or so) in series with a speaker from pin 7 to GND to hear the audio.

The code speed is set using the Frequency property of the CodeSpeed object.
Audio tone frequency is set using the Initial Frequency property of the PlayFrequency1 object.